Refactored client's pages and routes  #156
