+++ client/package.json
@@ -52,6 +52,7 @@
     "@quasar/cli": "^1.1.0",
     "@quasar/quasar-app-extension-qmediaplayer": "^1.1.2",
     "@quasar/quasar-app-extension-qpdfviewer": "^1.0.0-beta.8",
+    "@types/lodash": "^4.14.168",
     "@vue/eslint-config-standard": "^4.0.0",
     "babel-eslint": "^10.0.3",
     "eslint": "^6.8.0",
