+++ client/src/layouts/Base.vue
@@ -2,22 +2,27 @@
   <q-layout  view="hHh lpr fFf">
 
     <q-header style="min-width:320px" reveal bordered>
-      <NavBar @toggleDrawer="toggleDrawer"/>
+      <NavBar />
     </q-header>
 
     <q-page-container>
-      <q-banner v-if="isBackendDisconnected" inline-actions class="text-white bg-red">
+      <q-banner
+        v-if="isBackendDisconnected"
+        inline-actions
+        class="text-white bg-red"
+      >
         Disconnected from the backend. Start the server or change the port.
       </q-banner>
       <router-view />
     </q-page-container>
 
-    <q-toolbar bordered class="databrary-footer max-page-width bg-white text-black">
+    <q-toolbar
+      bordered
+      class="text-black"
+    >
       <div class="row full-width">
-        <div class="text-h6 col-12">
-          <div class="text-center text-subtitle1 col-12">
-            <small>Copyright &copy;2014-{{year}} Databrary</small>
-          </div>
+        <div class="col-12 text-h6 text-center text-subtitle1">
+          <small>Copyright &copy;2014-{{year}} Databrary</small>
         </div>
       </div>
     </q-toolbar>
@@ -36,7 +41,6 @@ export default {
   },
   data () {
     return {
-      leftDrawerOpen: false, // this.$q.platform.is.desktop,
       year: null
     }
   },
@@ -62,24 +66,7 @@ export default {
       } else {
         this.isBackendDisconnected = false
       }
-    },
-    toggleDrawer () {
-      this.leftDrawerOpen = !this.leftDrawerOpen
     }
   }
 }
 </script>
-
-<style>
-.q-layout__section--marginal {
-  background-color:#e0e0e0;
-}
-.databrary-footer {
-  width: 100%;
-}
-
-.max-page-width {
-  max-width: 1200px !important;
-  margin: auto;
-}
-</style>
