--- client/src/layouts/Project.vue
@@ -1,3 +1,0 @@
-<template>
-  <router-view />
-</template>
