--- client/src/router/project.js
@@ -1,9 +1,0 @@
-export default function routes (store) {
-  return [
-    {
-      name: 'projectLanding',
-      path: ':projectId',
-      component: () => import('pages/project/Landing2.vue')
-    }
-  ]
-}
